SCSI: refactor bus to be controller-agnostic
