+++ core/src/mac/compact/bus.rs
@@ -247,7 +247,7 @@ where
             .unwrap_or_else(|_| null_audio_sink());
         self.audio.set_sink(sink);
 
-        self.scsi.set_audio_provider(provider)
+        self.scsi.bus_mut().set_audio_provider(provider)
     }
 
     fn soundbuf(&mut self) -> &mut [u8] {
