+++ core/src/mac/scsi/controller.rs
@@ -1,8 +1,5 @@
 //! NCR5380 SCSI controller
 
-use std::collections::VecDeque;
-use std::path::{Path, PathBuf};
-
 use anyhow::{Result, bail};
 
 use log::*;
@@ -11,26 +8,13 @@ use num_derive::ToPrimitive;
 use num_traits::FromPrimitive;
 use proc_bitfield::bitfield;
 use serde::{Deserialize, Serialize};
-use serde_big_array::BigArray;
 
 use crate::bus::{Address, BusMember};
 use crate::dbgprop_byte;
 use crate::debuggable::Debuggable;
 use crate::emulator::EmuContext;
-use crate::mac::scsi::STATUS_GOOD;
-use crate::mac::scsi::ScsiCmdResult;
-use crate::mac::scsi::cdrom::ScsiTargetCdrom;
-use crate::mac::scsi::disk::ScsiTargetDisk;
-use crate::mac::scsi::disk_image::DiskImage;
-#[cfg(feature = "ethernet")]
-use crate::mac::scsi::ethernet::ScsiTargetEthernet;
-#[cfg(feature = "printer")]
-use crate::mac::scsi::printer::ScsiTargetPrinter;
+use crate::mac::scsi::bus::{CmdOutcome, ScsiBus};
 use crate::mac::scsi::scsi_cmd_len;
-use crate::mac::scsi::target::ScsiTarget;
-use crate::mac::scsi::target::ScsiTargetType;
-use crate::mac::scsi::toolbox::BlueSCSI;
-use crate::renderer::AudioProvider;
 use crate::tickable::{Tickable, Ticks};
 use crate::types::LatchingEvent;
 
@@ -202,7 +186,6 @@ pub struct ScsiController {
     reg_odr: u8,
     reg_bsr: NcrRegBsr,
     reg_selen: u8,
-    status: u8,
 
     /// DMA has been armed (Start DMA Send / Target Receive / Initiator
     /// Receive register written). Gates DRQ and the phase-mismatch IRQ
@@ -211,36 +194,14 @@ pub struct ScsiController {
     /// direction has been armed.
     dma_armed: bool,
 
-    /// Selected SCSI ID
-    sel_id: usize,
-
     /// Selected with attention
     sel_atn: bool,
 
-    /// Command buffer
-    cmdbuf: Vec<u8>,
-
-    /// Active command length
-    cmdlen: usize,
-
-    /// DataOut phase length
-    dataout_len: usize,
-
-    /// Response buffer
-    responsebuf: VecDeque<u8>,
-
-    /// Attached targets
-    #[serde(with = "BigArray")]
-    pub(crate) targets: [Option<Box<dyn ScsiTarget>>; Self::MAX_TARGETS],
+    /// The bus and its targets
+    pub(crate) bus: ScsiBus,
 
     set_req: LatchingEvent,
 
-    #[serde(skip)]
-    toolbox: BlueSCSI,
-    scsi_debug: bool,
-
-    #[serde(skip)]
-    scsi_trace_cdb: bool,
     #[serde(skip)]
     scsi_trace_phase: bool,
     #[serde(skip)]
@@ -248,62 +209,8 @@ pub struct ScsiController {
 }
 
 impl ScsiController {
-    pub const MAX_TARGETS: usize = 7;
-
     pub fn get_irq(&self) -> bool {
         self.reg_bsr.irq()
-    }
-
-    /// Returns the capacity of a target or None if detached or no media
-    pub fn get_disk_capacity(&self, id: usize) -> Option<usize> {
-        self.targets[id].as_ref().and_then(|t| t.capacity())
-    }
-
-    /// Returns the length of image data to write to savestates or None if no data
-    #[cfg(feature = "savestates")]
-    pub fn get_savestate_img_len(&self, id: usize) -> Option<usize> {
-        self.targets[id]
-            .as_ref()
-            .and_then(|t| t.savestate_img_len())
-    }
-
-    /// Returns the image filename of a target or None if detached or no media
-    pub fn get_disk_imagefn(&self, id: usize) -> Option<&Path> {
-        self.targets[id].as_ref().and_then(|t| t.image_fn())
-    }
-
-    /// Gets the target type (if attached) of an ID
-    pub fn get_target_type(&self, id: usize) -> Option<ScsiTargetType> {
-        self.targets[id].as_ref().map(|t| t.target_type())
-    }
-
-    /// Returns the Toolbox-managed image folder of a folder-backed CD-ROM at the
-    /// given ID, or None if the ID has no folder-backed CD-ROM attached.
-    pub fn get_cdrom_folder(&self, id: usize) -> Option<&Path> {
-        self.targets[id].as_ref().and_then(|t| t.cdrom_image_dir())
-    }
-
-    /// Builds the BlueSCSI Toolbox device-type map (one byte per SCSI ID) returned
-    /// by the LIST_DEVICES metadata subcommand. 0xFF means no device on that ID.
-    fn toolbox_device_map(&self) -> [u8; 8] {
-        let mut map = [0xFFu8; 8];
-        for (id, slot) in map.iter_mut().enumerate().take(Self::MAX_TARGETS) {
-            if let Some(target) = self.targets[id].as_ref() {
-                *slot = match target.target_type() {
-                    ScsiTargetType::Disk => 0x00,  // fixed disk
-                    ScsiTargetType::Cdrom => 0x02, // optical / CD-ROM
-                    #[cfg(feature = "ethernet")]
-                    ScsiTargetType::Ethernet => 0x06, // DaynaPORT
-                    #[cfg(feature = "printer")]
-                    ScsiTargetType::Printer => 0xFF, // not a Toolbox device type
-                };
-            }
-        }
-        map
-    }
-
-    pub fn set_shared_dir(&mut self, path: Option<PathBuf>) {
-        self.toolbox = BlueSCSI::new(path);
     }
 
     pub fn new() -> Self {
@@ -312,7 +219,6 @@ impl ScsiController {
                 .map(|v| v != "0" && !v.is_empty())
                 .unwrap_or(false)
         };
-        let scsi_trace_cdb = env_flag("SNOW_SCSI_TRACE_CDB");
         let scsi_trace_phase = env_flag("SNOW_SCSI_TRACE_PHASE");
         let scsi_trace_irq = env_flag("SNOW_SCSI_TRACE_IRQ");
         Self {
@@ -326,117 +232,12 @@ impl ScsiController {
             reg_odr: 0,
             reg_selen: 0,
             dma_armed: false,
-            sel_id: 0,
             sel_atn: false,
-            cmdbuf: vec![],
-            responsebuf: VecDeque::default(),
-            cmdlen: 0,
-            dataout_len: 0,
-            status: 0,
+            bus: ScsiBus::new(),
             set_req: Default::default(),
-            targets: Default::default(),
-            toolbox: BlueSCSI::default(),
-            scsi_debug: false,
-            scsi_trace_cdb,
             scsi_trace_phase,
             scsi_trace_irq,
         }
-    }
-
-    /// Loads a disk image (filename) and attaches a hard drive at the given SCSI ID
-    pub fn attach_hdd_at(&mut self, filename: &Path, scsi_id: usize) -> Result<()> {
-        if scsi_id >= Self::MAX_TARGETS {
-            bail!("SCSI ID out of range: {}", scsi_id);
-        }
-        if !Path::new(filename).exists() {
-            bail!("File {} does not exist", filename.to_string_lossy());
-        }
-        self.targets[scsi_id] = Some(Box::new(ScsiTargetDisk::load_disk(filename)?));
-        Ok(())
-    }
-
-    /// Attaches a disk backed by a custom disk image at the given SCSI ID.
-    pub(crate) fn attach_disk_image_at(
-        &mut self,
-        image: Box<dyn DiskImage>,
-        scsi_id: usize,
-    ) -> Result<()> {
-        if scsi_id >= Self::MAX_TARGETS {
-            bail!("SCSI ID out of range: {}", scsi_id);
-        }
-        self.targets[scsi_id] = Some(Box::new(ScsiTargetDisk::new(image)));
-        Ok(())
-    }
-
-    /// Attaches a CD-ROM drive at the given SCSI ID
-    pub fn attach_cdrom_at(
-        &mut self,
-        scsi_id: usize,
-        audio_provider: Option<&mut (dyn AudioProvider + '_)>,
-    ) {
-        self.targets[scsi_id] = Some(Box::new(ScsiTargetCdrom::new(audio_provider)));
-    }
-
-    /// Attaches a folder-backed ("Toolbox managed") CD-ROM drive at the given
-    /// SCSI ID. The folder's images are exposed to the guest through the BlueSCSI
-    /// Toolbox CD commands (LIST_CDS / SET_NEXT_CD / COUNT_CDS) and the first
-    /// image found is mounted automatically.
-    pub fn attach_cdrom_folder_at(
-        &mut self,
-        scsi_id: usize,
-        dir: PathBuf,
-        audio_provider: Option<&mut (dyn AudioProvider + '_)>,
-    ) -> Result<()> {
-        if scsi_id >= Self::MAX_TARGETS {
-            bail!("SCSI ID out of range: {}", scsi_id);
-        }
-        let mut cdrom = ScsiTargetCdrom::new(audio_provider);
-        cdrom.set_image_dir(Some(dir));
-        if let Err(e) = cdrom.mount_first_image() {
-            log::warn!("Folder CD-ROM at SCSI ID {}: {:#}", scsi_id, e);
-        }
-        self.targets[scsi_id] = Some(Box::new(cdrom));
-        Ok(())
-    }
-
-    /// Inserts a CD-ROM with the custom disk image at the given SCSI ID.
-    pub fn insert_cdrom_image_at(
-        &mut self,
-        image: Box<dyn DiskImage>,
-        scsi_id: usize,
-    ) -> Result<()> {
-        if scsi_id >= Self::MAX_TARGETS {
-            bail!("SCSI ID out of range: {}", scsi_id);
-        }
-        let Some(target) = self.targets[scsi_id].as_mut() else {
-            bail!("No target attached at SCSI ID {}", scsi_id);
-        };
-        target.load_image(image)
-    }
-
-    /// Attaches an Ethernet adapter at the given SCSI ID
-    #[cfg(feature = "ethernet")]
-    pub fn attach_ethernet_at(&mut self, scsi_id: usize) {
-        self.targets[scsi_id] = Some(Box::new(ScsiTargetEthernet::default()));
-    }
-
-    /// Attaches a LaserWriter IISC printer at the given SCSI ID
-    #[cfg(feature = "printer")]
-    pub fn attach_printer_at(&mut self, scsi_id: usize, output_dir: std::path::PathBuf) {
-        self.targets[scsi_id] = Some(Box::new(ScsiTargetPrinter::new(output_dir)));
-    }
-
-    /// Detaches a target from the given SCSI ID
-    pub fn detach_target(&mut self, scsi_id: usize) {
-        self.targets[scsi_id] = None;
-    }
-
-    pub fn set_audio_provider(&mut self, provider: &mut dyn AudioProvider) -> Result<()> {
-        for t in self.targets.iter_mut().flatten() {
-            t.set_audio_provider(provider)?;
-        }
-
-        Ok(())
     }
 
     /// Translates a SCSI ID on the bus (bit position) to a numeric ID
@@ -487,7 +288,7 @@ impl ScsiController {
             return;
         }
         let id = Self::translate_id(target_bits).unwrap();
-        if self.targets[id].is_none() {
+        if self.bus.targets[id].is_none() {
             // No device present at this ID
             self.set_phase(ScsiBusPhase::Free);
             return;
@@ -500,7 +301,7 @@ impl ScsiController {
             }
             self.reg_bsr.set_irq(true);
         }
-        self.sel_id = id;
+        self.bus.sel_id = id;
         self.sel_atn = self.reg_icr.assert_atn();
         self.set_phase(ScsiBusPhase::Command);
     }
@@ -509,7 +310,7 @@ impl ScsiController {
         if self.scsi_trace_phase {
             debug!(
                 "SCSI phase: {:?} -> {:?} (id={}, atn={})",
-                self.busphase, phase, self.sel_id, self.sel_atn
+                self.busphase, phase, self.bus.sel_id, self.sel_atn
             );
         }
 
@@ -527,22 +328,22 @@ impl ScsiController {
                 self.reg_icr.set_aip(false);
             }
             ScsiBusPhase::Command => {
-                self.cmdbuf.clear();
-                self.responsebuf.clear();
+                self.bus.cmdbuf.clear();
+                self.bus.responsebuf.clear();
                 self.reg_csr.set_bsy(true);
                 self.reg_csr.set_cd(true);
                 self.reg_csr.set_msg(false);
                 self.assert_req();
             }
             ScsiBusPhase::DataIn => {
-                if self.responsebuf.is_empty() {
+                if self.bus.responsebuf.is_empty() {
                     return self.set_phase(ScsiBusPhase::Status);
                 }
                 self.reg_csr.set_bsy(true);
                 self.reg_csr.set_cd(false);
                 self.reg_csr.set_io(true);
                 self.reg_csr.set_msg(false);
-                self.reg_cdr = self.responsebuf.pop_front().unwrap();
+                self.reg_cdr = self.bus.responsebuf.pop_front().unwrap();
 
                 self.assert_req();
             }
@@ -560,7 +361,7 @@ impl ScsiController {
                 self.reg_csr.set_io(true);
 
                 self.reg_csr.set_msg(false);
-                self.reg_cdr = self.status;
+                self.reg_cdr = self.bus.status;
 
                 self.assert_req();
             }
@@ -589,85 +390,26 @@ impl ScsiController {
         }
     }
 
+    /// Runs the command in the buffer and moves the bus to the phase the
+    /// target asked for.
     fn cmd_run(&mut self, outdata: Option<&[u8]>) -> Result<()> {
-        let cmd = &self.cmdbuf;
-        let cmd_op = cmd[0];
-
-        if self.scsi_trace_cdb {
-            debug!(
-                "SCSI CDB id={} {:02X?} dataout={}B",
-                self.sel_id,
-                cmd.as_slice(),
-                outdata.map(|d| d.len()).unwrap_or(0)
-            );
-        }
-
-        // File-sharing/metadata toolbox commands (0xD0-0xD6, 0xD9) are bus-global
-        // and handled by the shared BlueSCSI helper. The per-device CD switching
-        // commands (LIST_CDS 0xD7 / SET_NEXT_CD 0xD8 / COUNT_CDS 0xDA) fall through
-        // to the addressed target, which owns the folder of images and its media.
-        let result = match cmd_op {
-            0xD0..=0xD6 | 0xD9 => {
-                let toolbox_devices = self.toolbox_device_map();
-                Ok(self.toolbox.handle_command(
-                    cmd,
-                    outdata,
-                    &mut self.scsi_debug,
-                    &toolbox_devices,
-                ))
-            }
-            _ => {
-                let Some(target) = self.targets[self.sel_id].as_mut() else {
-                    bail!("SCSI command to disconnected target ID {}", self.sel_id);
-                };
-                target.cmd(cmd, outdata)
-            }
-        };
-
-        if self.scsi_trace_cdb {
-            match &result {
-                Ok(ScsiCmdResult::Status(s)) => {
-                    debug!("SCSI CDB {:02X} -> Status({:02X})", cmd_op, s);
-                }
-                Ok(ScsiCmdResult::DataIn(d)) => {
-                    debug!("SCSI CDB {:02X} -> DataIn {}B", cmd_op, d.len());
-                }
-                Ok(ScsiCmdResult::DataOut(n)) => {
-                    debug!("SCSI CDB {:02X} -> DataOut {}B", cmd_op, n);
-                }
-                Err(e) => debug!("SCSI CDB {:02X} -> Err: {:#}", cmd_op, e),
-            }
-        }
-
-        match result {
-            Ok(ScsiCmdResult::Status(s)) => {
-                self.status = s;
-
-                self.set_phase(ScsiBusPhase::Status);
-            }
-            Ok(ScsiCmdResult::DataIn(data)) => {
-                self.status = STATUS_GOOD;
-                self.responsebuf = VecDeque::from(data);
-                self.set_phase(ScsiBusPhase::DataIn);
-            }
-            Ok(ScsiCmdResult::DataOut(len)) => {
-                self.dataout_len = len;
-                self.responsebuf.clear();
-                if len == 0 {
-                    // [SPC-3] 6.7:
-                    // "A parameter list length of zero specifies that the Data-Out Buffer shall be empty. This condition
-                    // shall not be considered as an error."
-                    if let Err(e) = self.cmd_run(Some(&[])) {
-                        log::error!("SCSI command run error: {:#}", e);
-                    }
-                } else {
-                    self.set_phase(ScsiBusPhase::DataOut);
-                }
-            }
-            Err(e) => return Err(e),
+        match self.bus.cmd_run(outdata)? {
+            CmdOutcome::Status => self.set_phase(ScsiBusPhase::Status),
+            CmdOutcome::DataIn => self.set_phase(ScsiBusPhase::DataIn),
+            CmdOutcome::DataOut(_) => self.set_phase(ScsiBusPhase::DataOut),
         }
 
         Ok(())
+    }
+
+    /// The bus and its targets
+    pub fn bus(&self) -> &ScsiBus {
+        &self.bus
+    }
+
+    /// The bus and its targets
+    pub fn bus_mut(&mut self) -> &mut ScsiBus {
+        &mut self.bus
     }
 
     pub fn get_drq(&self) -> bool {
@@ -700,10 +442,10 @@ impl ScsiController {
 
         // Legacy PIO DataOut path (byte buffered here, ACK handled via ICR).
         if self.busphase == ScsiBusPhase::DataOut && self.phase_match() {
-            self.responsebuf.push_back(val);
-            self.dataout_len -= 1;
-            if self.dataout_len == 0 {
-                let datavec = Vec::from_iter(self.responsebuf.iter().cloned());
+            self.bus.responsebuf.push_back(val);
+            self.bus.dataout_len -= 1;
+            if self.bus.dataout_len == 0 {
+                let datavec = Vec::from_iter(self.bus.responsebuf.iter().cloned());
                 if let Err(e) = self.cmd_run(Some(&datavec)) {
                     log::error!("SCSI command run error: {:#}", e);
                 }
@@ -741,12 +483,12 @@ impl ScsiController {
     fn deassert_ack(&mut self) {
         match self.busphase {
             ScsiBusPhase::DataOut => {
-                if self.dataout_len > 0 {
+                if self.bus.dataout_len > 0 {
                     self.assert_req();
-                    self.responsebuf.push_back(self.reg_odr);
-                    self.dataout_len -= 1;
-                    if self.dataout_len == 0 {
-                        let datavec = Vec::from_iter(self.responsebuf.iter().cloned());
+                    self.bus.responsebuf.push_back(self.reg_odr);
+                    self.bus.dataout_len -= 1;
+                    if self.bus.dataout_len == 0 {
+                        let datavec = Vec::from_iter(self.bus.responsebuf.iter().cloned());
                         if let Err(e) = self.cmd_run(Some(&datavec)) {
                             log::error!("SCSI command run error: {:#}", e);
                         }
@@ -757,16 +499,16 @@ impl ScsiController {
                 }
             }
             ScsiBusPhase::Command => {
-                if self.cmdbuf.is_empty() {
-                    self.cmdlen = scsi_cmd_len(self.reg_odr).unwrap_or_else(|| {
+                if self.bus.cmdbuf.is_empty() {
+                    self.bus.cmdlen = scsi_cmd_len(self.reg_odr).unwrap_or_else(|| {
                         log::error!("Cmd length unknown for {:02X}", self.reg_odr);
                         6
                     });
                 }
-                self.cmdbuf.push(self.reg_odr);
-                if self.cmdbuf.len() >= self.cmdlen {
+                self.bus.cmdbuf.push(self.reg_odr);
+                if self.bus.cmdbuf.len() >= self.bus.cmdlen {
                     if let Err(e) = self.cmd_run(None) {
-                        error!("SCSI command ({:02X}) error: {}", self.cmdbuf[0], e);
+                        error!("SCSI command ({:02X}) error: {}", self.bus.cmdbuf[0], e);
                     }
                 } else {
                     self.assert_req();
@@ -779,7 +521,7 @@ impl ScsiController {
                 self.set_phase(ScsiBusPhase::Free);
             }
             ScsiBusPhase::DataIn => {
-                if let Some(b) = self.responsebuf.pop_front() {
+                if let Some(b) = self.bus.responsebuf.pop_front() {
                     self.reg_cdr = b;
                     self.assert_req();
                 } else {
@@ -934,9 +676,7 @@ impl BusMember<Address> for ScsiController {
 
 impl Tickable<&dyn EmuContext> for ScsiController {
     fn tick(&mut self, ticks: Ticks, ctx: &dyn EmuContext) -> Result<Ticks> {
-        for target in self.targets.iter_mut().flatten() {
-            target.tick(ticks, ctx)?;
-        }
+        self.bus.tick_targets(ticks, ctx)?;
 
         Ok(ticks)
     }
@@ -950,7 +690,7 @@ impl Debuggable for ScsiController {
         };
 
         let mut targets = vec![];
-        for (id, o_t) in self.targets.iter().enumerate() {
+        for (id, o_t) in self.bus.targets.iter().enumerate() {
             if let Some(t) = o_t {
                 targets.push(dbgprop_nest!(
                     format!("ID #{} - {:?}", id, t.target_type()),
@@ -973,18 +713,18 @@ impl Debuggable for ScsiController {
                     dbgprop_byte!("ODR", self.reg_odr),
                     dbgprop_byte!("BSR", self.reg_bsr.0),
                     dbgprop_byte!("TCR", self.reg_tcr.0),
-                    dbgprop_byte!("Status", self.status),
+                    dbgprop_byte!("Status", self.bus.status),
                     dbgprop_bool!("DMA armed", self.dma_armed),
                 ]
             ),
             dbgprop_enum!("Bus phase", self.busphase),
-            dbgprop_udec!("Selected ID", self.sel_id),
+            dbgprop_udec!("Selected ID", self.bus.sel_id),
             dbgprop_bool!("Attention", self.sel_atn),
             dbgprop_header!("Buffers"),
-            dbgprop_udec!("Command buffer len", self.cmdbuf.len()),
-            dbgprop_udec!("Command length", self.cmdlen),
-            dbgprop_udec!("Response buffer len", self.responsebuf.len()),
-            dbgprop_udec!("Data out len", self.dataout_len),
+            dbgprop_udec!("Command buffer len", self.bus.cmdbuf.len()),
+            dbgprop_udec!("Command length", self.bus.cmdlen),
+            dbgprop_udec!("Response buffer len", self.bus.responsebuf.len()),
+            dbgprop_udec!("Data out len", self.bus.dataout_len),
         ]
     }
 }
